Moved subscription state, handleUpdate, handleDelete, toggleLoadingState into Dashboard component.
